update the script

+++ Scripts/CreateStructure.sql
@@ -43,7 +43,7 @@ GO
 
 CREATE TABLE InterChange -- Õ¾ÄÚ»»³Ë
 (
-	Inter_Change_Id INT PRIMARY KEY,
+	Inter_Change_Id BIGINT PRIMARY KEY,
 	From_Station_Line_Id INT FOREIGN KEY REFERENCES StationLine(Station_Line_Id),
 	To_Station_Line_Id INT FOREIGN KEY REFERENCES StationLine(Station_Line_Id),
 	Cost TIME,
@@ -52,7 +52,7 @@ GO
 
 CREATE TABLE OuterChange -- Õ¾Íâ»»³Ë
 (
-	Outer_Change_Id INT PRIMARY KEY,
+	Outer_Change_Id BIGINT PRIMARY KEY,
 	From_Station_Line_Id INT FOREIGN KEY REFERENCES StationLine(Station_Line_Id),
 	To_Station_Line_Id INT FOREIGN KEY REFERENCES StationLine(Station_Line_Id),
 	Cost TIME,
